Fix bug

+++ src/entity/yakudo_score.rs
@@ -4,9 +4,11 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "yakudo_scores")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i64,
+    pub id: i32,
     pub username: String,
+    #[sea_orm(column_type = "BigInteger")]
     pub tweet_id: u64,
+    #[sea_orm(column_type = "BigInteger")]
     pub retweet_id: u64,
     pub score: f64,
     pub date: chrono::DateTime<chrono::Local>,
